Apply permissions when deleting plots - closes #1791

+++ src/main/java/com/forgeessentials/economy/plots/command/CommandPlot.java
@@ -204,9 +204,13 @@ public class CommandPlot extends ParserCommandBase
 
     public static void parseDelete(CommandParserArgs arguments)
     {
-        Plot plot = getPlot(arguments.sender);
-        arguments.confirm("Plot \"%s\" has been deleted.", plot.getNameNotNull());
-        Plot.deletePlot(plot);
+        if (arguments.hasPermission(Plot.PERM_DELETE))
+        {
+            Plot plot = getPlot(arguments.sender);
+            arguments.confirm("Plot \"%s\" has been deleted.", plot.getNameNotNull());
+            Plot.deletePlot(plot);
+        }
+        else throw new TranslatedCommandException(FEPermissions.MSG_NO_COMMAND_PERM);
     }
 
     public static void parseClaim(CommandParserArgs arguments)
